rollout http changes to all pages
