+++ client/public/layout.html
@@ -60,20 +60,13 @@
     {% block global_script %}
     {% endblock %}
 
-    <script type="module" defer src='{{ auto_bundles["navbar"] }}'></script>
-    <script type="module" defer src='{{ auto_bundles["footer"] }}'></script>
-
     {% block scripts %}
     {% endblock %}
 </head>
 
 <body>
-    <div id="nav-content"></div>
-
     {% block content %}
     <div class="app-content" id="app-content"></div>
     {% endblock %}
-
-    <div id="footer-content"></div>
 </body>
 </html>
