+++ client/src/components/Footer.svelte
