--- client/src/components/footer/main.ts
@@ -1,9 +1,0 @@
-import NavBar from './Footer.svelte';
-
-const nav_bar = new NavBar({
-	target: document.getElementById("footer-content"),
-	props: {
-	}
-});
-
-export default nav_bar;
