+++ client/src/global.less
@@ -10,17 +10,6 @@ body {
 	padding-top: 5rem;
 	box-sizing: border-box;
 	font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Oxygen-Sans, Ubuntu, Cantarell, "Helvetica Neue", sans-serif;
-
-	> .app-content {
-		padding-left: 1rem;
-		padding-right: 1rem;
-		min-height: ~"calc(100vh - 14rem)";
-
-		@media (min-width: 1024px) {
-			padding-left: 15%;
-			padding-right: 15%;
-		}
-	}
 }
 
 input:disabled {
