+++ client/vite.config.ts
@@ -10,8 +10,6 @@ export default defineConfig({
   build: {
     rollupOptions: {
       input: {
-        navbar: resolve(__dirname, './src/components/navbar/main.ts'),
-        footer: resolve(__dirname, './src/components/footer/main.ts'),
         sign_up: resolve(__dirname, './src/components/sign_up/main.ts'),
         login: resolve(__dirname, './src/components/login/main.ts'),
         home: resolve(__dirname, './src/components/index/main.ts'),
